SSL Certificate Expiration Checker
This utility reads the hostname either via command line or via CSV
file to initiate SSL handshake to obtain the certificate. Upon
successful handshake, the certificate is parsed and expiration
time is computed.

It usues a workerpool of go routines to queue up the tasks such
that each of these SSL connections would be made independently
and result is notified to the console and file writers over
corresponding channels.